pyproject.toml
[project]
name = "ok_palette_lab"
version = "0.1.0"
description = "Smooth color maps using Oklch color space."
authors = [
    { name = "Kenta Kabashima", email = "kenta_program37@hotmail.co.jp" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = []

[project.optional-dependencies]
plotly = ["plotly>=6"]

[dependency-groups]
tools = [
    "pyyaml (>=6.0.3,<7.0.0)",
    "colour-science (>=0.4.7,<0.5.0)",
    "networkx (>=3.6.1,<4.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "plotly (>=6.5.1,<7.0.0)",
    "jupyterlab (>=4.5.1,<5.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "ipython (>=9.9.0,<10.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "noise (>=1.2.2,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
]
doc = [
    "sphinx (>=8.2.3,<9.0.0)",
    "sphinx-orange-book-theme (>=0.16.0,<0.17.0)",
    "myst-parser (>=4.0.1,<5.0.0)",
    "myst-nb (>=1.3.0,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "doc8 (>=2.0.0,<3.0.0)",
    "sphinx-autobuild (>=2025.8.25,<2026.0.0)",
    "types-toml (>=0.10.8.20240310,<0.11.0.0)",
]

[tool.poetry]
packages = [
    { include = "ok_palette_lab" },
    # ok_palette_lab_tools is used only for development.
    { include = "ok_palette_lab_tools", format = [] },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pylint]
disable = ["R0801", "W0511", "C0301", "R0913"]
